+++ supvisors/instancestatus.py
@@ -168,7 +168,7 @@ class SupvisorsInstanceStatus:
     def serial(self):
         """ Return a serializable form of the SupvisorsInstanceStatus. """
         payload = {'identifier': self.identifier,
-                   'node_name': self.supvisors_id.host_name,
+                   'node_name': self.supvisors_id.host_id,
                    'port': self.supvisors_id.http_port,
                    'statecode': self.state.value, 'statename': self.state.name,
                    'sequence_counter': self.sequence_counter,